+++ js/offer.js
@@ -24,9 +24,9 @@ function getRandomItem(array) {
 
 
 function getArrayStrings(arr) {
-
   const lenArr = random(0, arr.length - 1) || 1;
   const someArr = arr.slice(0, lenArr);
+
   return someArr;
 }
